Windows: Changed the array type from int to int64 when reading the sample data to create the data structures and during actual population synthesis ...

+++ popgen_win/synthesizer_algorithm/demo_parallel_nogqs.py
@@ -51,11 +51,11 @@ def configure_and_run(fileLoc, geo, varCorrDict):
 # Identifying the number of housing units in the disaggregate sample
 # Make Sure that the file is sorted by hhid
     dbc.execute('select hhid, serialno from hhld_sample order by hhid')
-    hhld_sample = numpy.asarray(dbc.fetchall(), int)
+    hhld_sample = numpy.asarray(dbc.fetchall(), numpy.int64)
     hhld_units = dbc.rowcount
 
     dbc.execute('select hhid, serialno, pnum, personuniqueid from person_sample order by hhid, pnum')
-    person_sample = numpy.asarray(dbc.fetchall(), int)
+    person_sample = numpy.asarray(dbc.fetchall(), numpy.int64)
 
     housing_sample = hhld_sample
     housing_units = hhld_units
